+++ Core/OrderHelper.php
@@ -14,6 +14,8 @@ use Exception;
 use OxidEsales\Eshop\Application\Model\Basket;
 use OxidEsales\Eshop\Application\Model\User;
 use OxidEsales\Eshop\Core\DatabaseProvider;
+use OxidEsales\Eshop\Core\Exception\DatabaseConnectionException;
+use OxidEsales\Eshop\Core\Exception\DatabaseErrorException;
 use OxidEsales\Eshop\Core\Exception\OutOfStockException;
 use OxidEsales\Eshop\Core\Field;
 use OxidEsales\Eshop\Core\Registry;
@@ -33,6 +35,8 @@ use Wirecard\PaymentSdk\Response\FailureResponse;
 use Wirecard\PaymentSdk\Response\FormInteractionResponse;
 use Wirecard\PaymentSdk\Response\InteractionResponse;
 use Wirecard\PaymentSdk\Response\Response;
+use Wirecard\PaymentSdk\Response\SuccessResponse;
+use Wirecard\PaymentSdk\Transaction\CreditCardTransaction;
 
 /**
  * Helper class to handle orders
@@ -41,7 +45,6 @@ use Wirecard\PaymentSdk\Response\Response;
  */
 class OrderHelper
 {
-
     const PAY_ERROR_VARIABLE = 'payerror';
     const PAY_ERROR_ID = '-102';
     const PAY_ERROR_TEXT_VARIABLE = 'payerrortext';
@@ -136,6 +139,25 @@ class OrderHelper
 
         if ($oResponse instanceof InteractionResponse) {
             return self::_handleInteractionResponse($oResponse);
+        }
+
+        self::_onSuccessResponse($oResponse, $oBackendService, $oOrder);
+    }
+
+    /**
+     * @param Response       $oResponse
+     * @param BackendService $oBackendService
+     * @param Order          $oOrder
+     *
+     * @throws Exception
+     *
+     * @since 1.0.0
+     */
+    private static function _onSuccessResponse($oResponse, $oBackendService, $oOrder)
+    {
+        if (!is_null($oBackendService) && $oResponse instanceof SuccessResponse) {
+            self::_handleCreditCard($oResponse, $oBackendService, $oOrder);
+            ResponseHandler::onSuccessResponse($oResponse, $oBackendService, $oOrder);
         }
     }
 
@@ -494,5 +516,52 @@ class OrderHelper
             'country_id' => $oUser->oxuser__oxcountryid->value,
             'state_id' => $oUser->oxuser__oxstateid->value,
         ];
+    }
+
+    /**
+     * Handle credit card
+     *
+     * @param SuccessResponse $oResponse
+     * @param BackendService  $oBackendService
+     * @param Order           $oOrder
+     *
+     * @return void
+     *
+     * @since 1.3.0
+     */
+    private static function _handleCreditCard($oResponse, $oBackendService, $oOrder)
+    {
+        $oLogger = Registry::getLogger();
+
+        if (!self::_shouldSaveCard($oResponse, $oOrder)) {
+            return;
+        }
+
+        try {
+            $aTransaction = $oBackendService->getTransactionByTransactionId(
+                $oResponse->getTransactionId(),
+                CreditCardTransaction::NAME
+            );
+
+            $oVault = new Vault();
+            $oVault->saveCard($oResponse, $aTransaction['payment']['card'], $oOrder);
+        } catch (DatabaseConnectionException $oExc) {
+            $oLogger->error("Cannot save card", [$oExc]);
+        } catch (DatabaseErrorException $oExc) {
+            $oLogger->error("Cannot save card", [$oExc]);
+        }
+    }
+
+    /**
+     * @param SuccessResponse $oResponse
+     * @param Order           $oOrder
+     *
+     * @return bool
+     *
+     * @since 1.3.0
+     */
+    private static function _shouldSaveCard($oResponse, $oOrder)
+    {
+        return $oResponse->getPaymentMethod() === CreditCardTransaction::NAME && $oOrder->getUser()->hasAccount();
     }
 }
